pyproject.toml
[project]
name = "llm-council"
version = "0.1.0"
description = "Your LLM Council"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.0",
    "pypdf>=4.0.0",
    "reportlab>=4.0.0",
    "pymupdf>=1.24.0",
    "pytesseract>=0.3.10",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
